move orphaner and helpers from lib/actions to lib/api for consistency
