+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/toolkit.ts
@@ -51,8 +51,8 @@ import { appendObject, prepareDiff } from '../actions/diff/private';
 import type { DriftOptions, DriftResult } from '../actions/drift';
 import { type ListOptions } from '../actions/list';
 import type { OrphanOptions } from '../actions/orphan';
-import { ResourceOrphaner } from '../actions/orphan/orphaner';
-import { parseAndValidateConstructPaths } from '../actions/orphan/private/helpers';
+import { ResourceOrphaner } from '../api/orphan/orphaner';
+import { parseAndValidateConstructPaths } from '../api/orphan/private/helpers';
 import type { PublishAssetsOptions, PublishAssetsResult } from '../actions/publish-assets';
 import type { RefactorOptions } from '../actions/refactor';
 import { type RollbackOptions } from '../actions/rollback';
